Move modules in sub namespaces into folders named them
* Modify Readme.md
