+++ Funcy/Patterns/Case.cs
