+++ Funcy/Patterns/IExtractor.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace Funcy
+namespace Funcy.Patterns
 {
     public interface IExtractor<T>
     {
