+++ Funcy/Patterns/Matcher.cs
