+++ Funcy/Patterns/Pattern.cs
